DM-14: Develop Hip Switch for business-registration Renaming model.des package to model.api
